Refactor package and class name 각각의 클래스와 패키지 별로 해당 역할에 맞는 이름으로 변경
