+++ src/main/java/com/codessquad/qna/controller/QuestionController.java
@@ -1,7 +1,7 @@
 package com.codessquad.qna.controller;
 
 import com.codessquad.qna.model.Question;
-import com.codessquad.qna.repository.Questions;
+import com.codessquad.qna.service.QuestionService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 public class QuestionController {
 
     private Logger logger = LoggerFactory.getLogger(UserController.class);
-    private Questions questions = new Questions();
+    private QuestionService questions = new QuestionService();
 
     @GetMapping("/")
     public String viewMain(Model model) {
