+++ src/main/resources/templates/index.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="kr">
 <head>
-    {{> component/header}}
+    {{> partial/header}}
 </head>
 <body>
-    {{> component/navigation}}
+    {{> partial/navigation}}
     <div class="container" id="main">
        <div class="col-md-12 col-sm-12 col-lg-10 col-lg-offset-1">
           <div class="panel panel-default qna-list">
@@ -127,6 +127,6 @@
   </div>
 </div>
 -->
-    {{> component/script}}
+    {{> partial/script}}
 </body>
 </html>
