+++ src/main/resources/templates/partial/navigation.html
