+++ src/main/resources/templates/partial/script.html
