+++ src/main/resources/templates/qna/form.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="kr">
 <head>
-    {{> component/header}}
+    {{> partial/header}}
 </head>
 <body>
-    {{> component/navigation}}
+    {{> partial/navigation}}
     <div class="container" id="main">
        <div class="col-md-12 col-sm-12 col-lg-10 col-lg-offset-1">
           <div class="panel panel-default content-main">
@@ -27,6 +27,6 @@
             </div>
         </div>
     </div>
-    {{> component/script}}
+    {{> partial/script}}
 </body>
 </html>
