+++ src/main/resources/templates/qna/show.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="kr">
 <head>
-    {{> component/header}}
+    {{> partial/header}}
 </head>
 <body>
-    {{> component/navigation}}
+    {{> partial/navigation}}
     <div class="container" id="main">
         <div class="col-md-12 col-sm-12 col-lg-12">
             <div class="panel panel-default">
@@ -153,6 +153,6 @@
             </div>
         </article>
     </script>
-    {{> component/script}}
+    {{> partial/script}}
 </body>
 </html>
