+++ src/main/resources/templates/user/form.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="kr">
 <head>
-    {{> component/header}}
+    {{> partial/header}}
 </head>
 <body>
-    {{> component/navigation}}
+    {{> partial/navigation}}
     <div class="container" id="main">
        <div class="col-md-6 col-md-offset-3">
           <div class="panel panel-default content-main">
@@ -31,6 +31,6 @@
             </div>
         </div>
     </div>
-    {{> component/script}}
+    {{> partial/script}}
 </body>
 </html>
