+++ src/main/resources/templates/user/updateForm.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="kr">
 <head>
-    {{> component/header}}
+    {{> partial/header}}
 </head>
 <body>
-    {{> component/navigation}}
+    {{> partial/navigation}}
     <div class="container" id="main">
        <div class="col-md-6 col-md-offset-3">
           <div class="panel panel-default content-main">
@@ -19,7 +19,7 @@
                       <input type="password" class="form-control" id="password" name="password" placeholder="Password">
                   </div>
                   <div class="form-group">
-                      <label for="newPassword">변경 할 비밀번호</label>
+                      <label for="newPassword">새로운 비밀번호</label>
                       <input type="password" class="form-control" id="newPassword" name="newPassword" placeholder="Password">
                   </div>
                   <div class="form-group">
@@ -36,6 +36,6 @@
             </div>
         </div>
     </div>
-    {{> component/script}}
+    {{> partial/script}}
 </body>
 </html>
